embedType QP and pr comments
