+++ src/index.ts
@@ -7,13 +7,20 @@ export {
 export {
   embedTemplateEditorPage,
   type EmbedTemplateEditorPageOptions,
+  TEMPLATE_EDITOR_FEATURES_DEFAULT,
+  TEMPLATE_EDITOR_FEATURES_ALL,
+  TEMPLATE_EDITOR_FEATURES_NONE,
+  TEMPLATE_EDITOR_FEATURES_ENABLE_BARCODE,
 } from "./calls/embedTemplateEditorPage.ts";
 export {makeSdk, type SdkOptions} from "./calls/makeSdk.ts";
 export {
   type TemplateEmbedFeatures,
   type TemplateEditorFeatures,
-  TEMPLATE_EDITOR_FEATURES_DEFAULT,
-  TEMPLATE_EDITOR_FEATURES_ALL,
-  TEMPLATE_EDITOR_FEATURES_NONE,
-  TEMPLATE_EDITOR_FEATURES_ENABLE_BARCODE,
 } from "./helpers/embedFeatures.ts";
+export {
+  type AppearanceConfig,
+  type FontSource,
+  type CssFontSource,
+  type CustomFontSource,
+} from "./helpers/appearance.ts";
+export {type SdkTokenPayload} from "./helpers/token.ts";
