+++ src/utils/iframe.ts
@@ -1,10 +1,14 @@
-import type {TemplateEmbedFeatures} from "../helpers/embedFeatures.ts";
+import type {
+  EmbedType,
+  TemplateEmbedFeatures,
+} from "../helpers/embedFeatures.ts";
 import type {FontSource, AppearanceConfig} from "../helpers/appearance";
 
-interface CreateIframeOptions {
+export interface CreateIframeOptions {
   token: string;
   path: string;
   pathHash: string;
+  embedType: EmbedType;
   config: {
     features?: TemplateEmbedFeatures | undefined;
     fonts?: FontSource[] | undefined;
@@ -15,7 +19,7 @@ interface CreateIframeOptions {
 export function createEmbedIframe(
   options: CreateIframeOptions,
 ): HTMLIFrameElement {
-  const {path, token, pathHash, config} = options;
+  const {path, token, pathHash, embedType, config} = options;
 
   const iframe = document.createElement("iframe");
 
@@ -26,6 +30,7 @@ export function createEmbedIframe(
   const iframeUrl = new URL(path);
 
   iframeUrl.searchParams.set("token", token);
+  iframeUrl.searchParams.set("embedType", embedType);
   iframeUrl.searchParams.set("config", JSON.stringify(config));
 
   iframeUrl.hash = pathHash;
